+++ assets/src/scripts/components/graph/view/water-levels.spec.js
@@ -1,3 +1,4 @@
+import { scaleLinear } from 'd3-scale';
 import { select, selectAll } from 'd3-selection';
 
 import { default as drawWaterLevels, drawDataLine } from './water-levels';
@@ -29,10 +30,7 @@ describe('graph component water levels', () => {
                         dateTime: new Date('2010-10-10'),
                         value: 10
                     }],
-                    classes: {
-                        approved: true,
-                        provisional: false
-                    }
+                    class: 'approved'
                 },
                 xScale: () => 1,
                 yScale: () => 1
@@ -56,13 +54,10 @@ describe('graph component water levels', () => {
                         dateTime: new Date('2010-10-11'),
                         value: 11
                     }],
-                    classes: {
-                        approved: true,
-                        provisional: false
-                    }
+                    class: 'approved'
                 },
-                xScale: () => 1,
-                yScale: () => 1
+                xScale: scaleLinear(),
+                yScale: scaleLinear()
             });
 
             const segments = selectAll('.line-segment');
@@ -84,22 +79,26 @@ describe('graph component water levels', () => {
                     dateTime: new Date('2010-10-11'),
                     value: 11
                 }],
-                classes: {
-                    approved: true,
-                    provisional: false
-                }
+                class: 'approved'
             }, {
                 points: [{
                     dateTime: new Date('2010-10-10'),
                     value: 10
                 }],
-                classes: {
-                    approved: true,
-                    provisional: false
-                }
+                class: 'approved'
             }],
-            xScale: () => 1,
-            yScale: () => 1
+            chartPoints: [{
+                dateTime: new Date('2010-10-10'),
+                value: 10
+            }, {
+                dateTime: new Date('2010-10-11'),
+                value: 11
+            }, {
+                dateTime: new Date('2010-10-10'),
+                value: 10
+            }],
+            xScale: scaleLinear(),
+            yScale: scaleLinear()
         });
         const group = select('g#ts-group');
         expect(group).not.toBe(null);