+++ CHANGELOG.md
@@ -13,6 +13,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added water quality table to site location view
 - Added water levels graph to site location page
 - Added brush/pan chart under main chart
-- Add legend to chart
+- Add legend to chart when there are classed points
 - Line gaps drawn on gap of 6 months or more
 - y-axis label set to depth to water, feet below land surface